+++ ui-library-call-readiness/src/App.tsx
@@ -4,8 +4,8 @@ import {
   DEFAULT_COMPONENT_ICONS
 } from '@azure/communication-react';
 import { useState } from 'react';
-import { TestComplete } from './components/TestComplete';
-import CallReadinessChecks from './CallReadinessChecks';
+import { TestComplete } from './TestComplete';
+import CallReadinessChecks from './call-readiness/CallReadinessCheckComponent';
 
 // Initializing and registering icons should only be done once per app.
 initializeIcons();
