+++ ui-library-call-readiness/src/call-readiness/UnsupportedBrowserPrompt.tsx
@@ -1,7 +1,7 @@
 import { UnsupportedBrowser } from '@azure/communication-react';
 import { Modal } from '@fluentui/react';
 
-export const BrowserUnsupportedModal = (props: { isOpen: boolean }): JSX.Element => (
+export const BrowserUnsupportedPrompt = (props: { isOpen: boolean }): JSX.Element => (
   <Modal isOpen={props.isOpen}>
     <UnsupportedBrowser
       onTroubleshootingClick={() => alert('This callback should be used to take the user to further troubleshooting')}
